Fix the warning in GenericPlatformBackgroundHttpRequest.cpp: Expression 'ByteDifference >= 0' is always true. Unsigned type value is always >= 0.
[REVIEW] [at]michael.kirzinger [at]rafa.lecina [at]michael.atchison
#rb [at]michael.kirzinger [at]rafa.lecina

[CL 29416138 by lorry li in ue5-main branch]

+++ Engine/Source/Runtime/Online/BackgroundHTTP/Public/GenericPlatform/GenericPlatformBackgroundHttpRequest.h
@@ -67,7 +67,7 @@ protected:
 		int32 MaxRetries;
 
 		//How many bytes we had last time we sent progress updates
-		int32 LastProgressUpdateBytes;
+		uint64 LastProgressUpdateBytes;
 	};
 
 	TUniquePtr<FGenericPlatformBackgroundHttpWrapper> RequestWrapper;
